Update code to actually do things
Built out findChannel function to find the channel.
Added logic to determine if findChannel successful, and if not to insert current channel and the character object (as init_table). If successful, the function will pull the init_table, push the char and roll to the init_table, and then update the record in the db collection.